# services/api/routers/runs.py
"""Run 生命周期控制面路由。

流程：HTTP → DTO → use case（RunOrchestrationService）→ Domain / Port。
Timeline/事件端点见 run_events.py（SSE + JSON replay 双模式）。
"""

from __future__ import annotations

from dataclasses import dataclass

from fastapi import APIRouter, Request

from packages.application.ports import CatalogSnapshot, PreflightContext, ProjectSettings
from packages.application.run_orchestration.commands import CancelRunCommand, StartRunCommand
from packages.domain.core import ID, Digest
from packages.domain.protocols import ProtocolDefinition
from packages.domain.run import ResearchRun
from packages.domain.state_base import InvalidTransitionError
from services.api.catalog import (
    load_catalog_snapshot,
    load_project_settings,
    load_protocol_definition,
)
from services.api.composition import ApiDeps
from services.api.deps import get_deps
from services.api.dto.runs import RunDetailDto, RunStartDto, TaskDto
from services.api.errors import ApiError
from services.api.routers.run_events import events_of

router = APIRouter(prefix="/runs", tags=["runs"])
projects_router = APIRouter(tags=["runs"])


def _run_state_dto(deps: ApiDeps, run_id: str) -> RunDetailDto:
    run = deps.run_registry.get(run_id)
    if run is None:
        raise ApiError(404, "Not Found", f"run not found: {run_id}")
    manifest = str(run.manifest_digest) if run.manifest_digest else None
    return RunDetailDto(
        id=run.id.value,
        project_id=run.project_id,
        protocol_id=run.protocol_id,
        state=run.state,
        manifest_digest=manifest,
        created_at=run.created_at.value.isoformat(),
        updated_at=run.updated_at.value.isoformat(),
    )


def _frozen_manifest_digest_of(deps: ApiDeps, run_id: str) -> str | None:
    """从事件投影恢复 MANIFEST_FROZEN digest（execution 失败时用）。"""
    if deps.projection is None:
        return None
    for envelope in events_of(deps.projection, run_id):
        if envelope.event_type.value == "manifest.frozen":
            digest = envelope.payload.get("digest")
            if isinstance(digest, str):
                return digest
    return None


def _run_from_execution(
    deps: ApiDeps,
    run_id: ID,
    project_id: str,
    protocol_id: str,
    inputs: ExecutionInputs,
) -> ResearchRun:
    """执行链结果 → run 实体（执行期 ValueError 收敛 FAILED + 保留 frozen digest）。"""
    if deps.runs is None:
        raise ApiError(503, "Run Orchestration Unavailable", "run service not configured")
    try:
        outcome = deps.runs.start_run(
            inputs.protocol, inputs.catalog, inputs.project, inputs.preflight, inputs.command
        )
        return ResearchRun(
            id=run_id,
            project_id=project_id,
            protocol_id=protocol_id,
            state=outcome.state,
            manifest_digest=Digest.parse(outcome.manifest_digest)
            if outcome.manifest_digest
            else None,
        )
    except ValueError:
        frozen_digest = _frozen_manifest_digest_of(deps, run_id.value)
        return ResearchRun(
            id=run_id,
            project_id=project_id,
            protocol_id=protocol_id,
            state="FAILED",
            manifest_digest=Digest.parse(frozen_digest) if frozen_digest else None,
        )


@dataclass(frozen=True, slots=True)
class ExecutionInputs:
    """start_run 执行输入聚合（参数对象，规避参数爆发）。"""

    protocol: ProtocolDefinition
    catalog: CatalogSnapshot
    project: ProjectSettings
    preflight: PreflightContext
    command: StartRunCommand


def _execution_inputs(
    deps: ApiDeps,
    protocol_path: str,
    run_id: ID,
    trace_id: str | None,
) -> ExecutionInputs:
    """加载协议/目录/项目并构建命令（override 时 catalog/project 与 preflight 同源）。"""
    protocol = load_protocol_definition(protocol_path)
    catalog = load_catalog_snapshot()
    project = load_project_settings()
    preflight = deps.preflight_override
    if preflight is not None:
        catalog = preflight.catalog
        project = preflight.project
    else:
        preflight = PreflightContext(catalog=catalog, project=project)
    command = StartRunCommand(
        project_id=project.project_id,
        protocol_id=protocol.id,
        run_id=run_id,
        trace_id=trace_id or f"api-{run_id.value}",
    )
    return ExecutionInputs(protocol, catalog, project, preflight, command)


@projects_router.post("/projects/{project_id}/runs", response_model=RunDetailDto)
async def start_run(project_id: str, payload: RunStartDto, request: Request) -> RunDetailDto:
    """启动一次 Research Run（StartRunCommand → 正式编排链）。"""
    del project_id
    deps: ApiDeps = get_deps(request)
    run_id = ID.generate()
    inputs = _execution_inputs(deps, payload.protocol_path, run_id, payload.trace_id)
    run = _run_from_execution(deps, run_id, inputs.project.project_id, inputs.protocol.id, inputs)
    deps.run_registry[run_id.value] = run
    return _run_state_dto(deps, run_id.value)


@router.get("/{run_id}", response_model=RunDetailDto)
async def get_run(run_id: str, request: Request) -> RunDetailDto:
    """Run 状态视图（canonical state；刷新后从 API 恢复）。"""
    deps: ApiDeps = get_deps(request)
    return _run_state_dto(deps, run_id)


@router.post("/{run_id}/cancel", response_model=RunDetailDto)
async def cancel_run(run_id: str, request: Request) -> RunDetailDto:
    """协作式取消（幂等）；非法状态机迁移 → 409。"""
    deps: ApiDeps = get_deps(request)
    if deps.runs is None:
        raise ApiError(503, "Run Orchestration Unavailable", "run service not configured")
    run = deps.run_registry.get(run_id)
    if run is None:
        raise ApiError(404, "Not Found", f"run not found: {run_id}")
    try:
        deps.runs.cancel_run(CancelRunCommand(run_id=ID(run_id)))
        cancelled = run.transition("CANCEL")
        deps.run_registry[run_id] = cancelled
    except InvalidTransitionError as exc:
        raise ApiError(409, "Invalid Transition", str(exc)) from exc
    except ValueError as exc:
        raise ApiError(409, "Invalid Transition", str(exc)) from exc
    return _run_state_dto(deps, run_id)


@router.get("/{run_id}/tasks", response_model=list[TaskDto])
async def list_run_tasks(run_id: str, request: Request) -> list[TaskDto]:
    """run 任务投影（canonical state 读取，非审计事件）。"""
    deps: ApiDeps = get_deps(request)
    if run_id not in deps.run_registry:
        raise ApiError(404, "Not Found", f"run not found: {run_id}")
    if deps.projection is None:
        raise ApiError(503, "Projection Unavailable", "run projection not configured")
    rows = deps.projection.list_tasks(run_id)
    return [
        TaskDto(
            task_id=task.id.value,
            contract_id=contract.id,
            agent_id=task.assigned_agent_id,
            status=task.status,
            attempt=task.attempt,
        )
        for task, contract in rows
    ]